refactoring

+++ src/main/java/team1/myshop/web/ItemService.java
@@ -53,7 +53,6 @@ public class ItemService extends ServiceBase {
     }
 
     @POST
-    @Path("/new")
     @Produces(MediaType.APPLICATION_JSON)
     @Consumes(MediaType.APPLICATION_JSON)
     public Item createItem(String itemString, @Context HttpServletRequest request,
@@ -62,17 +61,23 @@ public class ItemService extends ServiceBase {
         this.initialize();
 
         // check user rights
-        auth.ensureUserRight(request, response, UserRights.CAN_CREATE_ITEM);
+        if (!auth.userHasRight(request, UserRights.CAN_CREATE_ITEM)) {
+            logger.info("User tried to create an item, but did not have the right to");
+            http.cancelRequest(response, SC_UNAUTHORIZED);
+            return null;
+        }
 
         Item item = JsonParser.parse(itemString, Item.class);
         if (item == null) {
             http.cancelRequest(response, SC_BAD_REQUEST);
+            return null;
         }
 
         assert item != null;
         data.model.Item it = dh.createItem(item.title, item.description, item.categoryID, item.authorID);
         if (it == null) {
             http.cancelRequest(response, SC_INTERNAL_SERVER_ERROR);
+            return null;
         }
 
         assert it != null;
@@ -92,18 +97,24 @@ public class ItemService extends ServiceBase {
         this.initialize();
 
         // check user rights
-        auth.ensureUserRight(request, response, UserRights.CAN_EDIT_ITEM);
+        if (!auth.userHasRight(request, UserRights.CAN_EDIT_ITEM)) {
+            logger.info("User tried to create an item, but did not have the right to");
+            http.cancelRequest(response, SC_UNAUTHORIZED);
+            return null;
+        }
 
         // parse body data
         Item it = JsonParser.parse(itemString, Item.class);
         if (it == null) {
             http.cancelRequest(response, SC_BAD_REQUEST);
+            return null;
         }
 
         // Parameter müssen übereinstimmen
         assert it != null;
         if (it.id != itemID) {
             http.cancelRequest(response, SC_BAD_REQUEST);
+            return null;
         }
 
         data.model.Item item = dh.changeItem(itemID, it.title, it.description, it.categoryID);
@@ -120,7 +131,12 @@ public class ItemService extends ServiceBase {
         this.initialize();
 
         // check user rights
-        auth.ensureUserRight(request, response, UserRights.CAN_DELETE_ITEM);
+        // check user rights
+        if (!auth.userHasRight(request, UserRights.CAN_DELETE_ITEM)) {
+            logger.info("User tried to delete an item, but did not have the right to");
+            http.cancelRequest(response, SC_UNAUTHORIZED);
+            return;
+        }
 
         // delete item
         dh.deleteItem(item);
